requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
certifi==2021.10.8
charset-normalizer==2.0.10; python_version >= '3'
click==8.0.3; python_version >= '3.6'
colorama==0.4.4; platform_system == 'Windows'
decorator==5.1.1; python_version >= '3.5'
elastic-transport==8.0.0a7; python_version >= '3.6'
elasticsearch==8.0.0a4
flask==2.0.2
idna==3.3; python_version >= '3'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
python-dotenv==0.19.2
requests==2.27.1
retry==0.9.2
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.2; python_version >= '3.6'
